requirements.txt
# Hundtracker Backend - Python dependencies
# Core
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# MQTT
aiomqtt>=2.0.0
paho-mqtt==2.1.0

# Config & validation
pydantic==2.10.2
pydantic-settings==2.6.1

# WebSocket & async
python-multipart==0.0.17
